+++ .storybook/main.js
@@ -3,18 +3,29 @@ const path = require('path');
 const maxAssetSize = 250 * 1024;
 
 module.exports = {
-  stories: ['../src/**/*.stories.tsx'],
+  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-essentials',
-    '@storybook/addon-links',
+    '@storybook/addon-interactions',
     '@storybook/addon-postcss',
     '@storybook/addon-storysource',
   ],
+  core: {
+    builder: 'webpack5',
+  },
+  framework: '@storybook/react',
   webpackFinal: async (config) => {
+    config.module.rules.unshift({
+      test: /\.scss$/i,
+      resourceQuery: /raw/,
+      type: 'asset/source',
+    });
+
     // SCSS ALL EXCEPT local
     config.module.rules.push({
       test: /\.module\.scss$/i,
+      resourceQuery: { not: [/raw/] },
       use: ['style-loader', 'css-loader', 'sass-loader'],
     });
 
@@ -22,6 +33,7 @@ module.exports = {
     config.module.rules.push({
       test: /\.scss$/i,
       exclude: /\.module\.scss$/i,
+      resourceQuery: { not: [/raw/] },
       use: ['style-loader', 'css-loader', 'sass-loader'],
       include: path.resolve(__dirname, '../'),
     });
@@ -46,4 +58,15 @@ module.exports = {
     return config;
   },
   staticDirs: ['../static'],
+  managerHead: (head) => `${head}
+  <script async src="https://www.googletagmanager.com/gtag/js?id=G-LXDCMHQK8V"></script>
+  <script>
+    window.dataLayer = window.dataLayer || [];
+    function gtag() {
+      dataLayer.push(arguments);
+    }
+    gtag('js', new Date());
+  
+    gtag('config', 'G-LXDCMHQK8V');
+  </script>`,
 };