+++ src/Components/Organisms/DataTable/Common/Cells/__test__/MultiSelectCell.test.tsx
@@ -41,6 +41,7 @@ describe('MultiSelectCell component', () => {
             <MultiSelectCell
               column={{
                 dataIndex: 'data',
+                editable: true,
                 numberOfItemLabel: 'numberOfItemLabel',
                 numberOfItemsLabel: 'numberOfItemsLabel',
                 options: () => [