+++ src/Components/Organisms/Navbar/index.ts
@@ -1,3 +1,0 @@
-export { default as Navbar } from './Navbar';
-export { default as NavItem } from './NavItem';
-export type { INavItemProps } from './NavItem';