+++ src/Components/Organisms/SideBar/index.ts
@@ -1,0 +1,5 @@
+export type { ISideBarItemProps } from './SideBarItem';
+export type { ISideBarSectionProps } from './SideBarSection';
+export { default as SideBar } from './SideBar';
+export { default as SideBarItem } from './SideBarItem';
+export { default as SideBarSection } from './SideBarSection';