+++ src/Components/Organisms/index.ts
@@ -2,5 +2,5 @@ export * from './ActionBar';
 export * from './DataTable';
 export * from './Filter';
 export * from './Form';
-export * from './Navbar';
-export * from './Sidebar';
+export * from './NavBar';
+export * from './SideBar';